+++ api-tests/relationships/filtering/nested.test.js
@@ -1,5 +1,6 @@
 const { Text, Relationship } = require('@keystonejs/fields');
-const { multiAdapterRunners, setupServer, graphqlRequest } = require('@keystonejs/test-utils');
+const { multiAdapterRunners, setupServer } = require('@keystonejs/test-utils');
+const { createItem, createItems } = require('@keystonejs/server-side-graphql-client');
 
 function setupKeystone(adapterName) {
   return setupServer({
@@ -31,20 +32,27 @@ multiAdapterRunners().map(({ runner, adapterName }) =>
     describe('relationship filtering', () => {
       test(
         'nested to-many relationships can be filtered',
-        runner(setupKeystone, async ({ keystone, create }) => {
-          const ids = [];
-
-          ids.push((await create('Post', { content: 'Hello world' })).id);
-          ids.push((await create('Post', { content: 'hi world' })).id);
-          ids.push((await create('Post', { content: 'Hello? Or hi?' })).id);
-
-          const user = await create('User', { posts: ids });
-
-          // Create a dummy user to make sure we're actually filtering it out
-          const user2 = await create('User', { posts: [ids[0]] });
-
-          const { data, errors } = await graphqlRequest({
+        runner(setupKeystone, async ({ keystone }) => {
+          const ids = await createItems({
             keystone,
+            listKey: 'Post',
+            items: [
+              { data: { content: 'Hello world' } },
+              { data: { content: 'hi world' } },
+              { data: { content: 'Hello? Or hi?' } },
+            ],
+          });
+
+          const [user, user2] = await createItems({
+            keystone,
+            listKey: 'User',
+            items: [
+              { data: { posts: { connect: ids } } },
+              { data: { posts: { connect: [ids[0]] } } }, // Create a dummy user to make sure we're actually filtering it out
+            ],
+          });
+
+          const { data, errors } = await keystone.executeGraphQL({
             query: `
         query {
           allUsers {
@@ -63,39 +71,45 @@ multiAdapterRunners().map(({ runner, adapterName }) =>
           expect(errors).toBe(undefined);
           expect(data).toHaveProperty('allUsers.0.posts');
           expect(data.allUsers).toHaveLength(2);
-          expect(data.allUsers[0].id).toEqual(user.id);
-          expect(data.allUsers[0].posts).toHaveLength(2);
-          expect(data.allUsers[0].posts.map(({ id }) => id).sort()).toEqual(
-            [ids[1], ids[2]].sort()
-          );
-          expect(data.allUsers[1].id).toEqual(user2.id);
-          expect(data.allUsers[1].posts).toHaveLength(0);
+          data.allUsers[0].posts = data.allUsers[0].posts.map(({ id }) => id).sort();
+          data.allUsers[1].posts = data.allUsers[1].posts.map(({ id }) => id).sort();
+          expect(data.allUsers).toContainEqual({
+            id: user.id,
+            posts: [ids[1].id, ids[2].id].sort(),
+          });
+          expect(data.allUsers).toContainEqual({ id: user2.id, posts: [] });
         })
       );
 
       test(
         'nested to-many relationships can be limited',
-        runner(setupKeystone, async ({ keystone, create }) => {
-          const ids = [];
-
-          ids.push((await create('Post', { content: 'A hello world' })).id);
-          ids.push((await create('Post', { content: 'hi world' })).id);
-          ids.push((await create('Post', { content: 'Hello? Or hi?' })).id);
-
-          const user = await create('User', { posts: ids });
-
-          // Create a dummy user to make sure we're actually filtering it out
-          const user2 = await create('User', { posts: [ids[0]] });
-
-          const { data, errors } = await graphqlRequest({
+        runner(setupKeystone, async ({ keystone }) => {
+          const ids = await createItems({
             keystone,
+            listKey: 'Post',
+            items: [
+              { data: { content: 'Hello world' } },
+              { data: { content: 'hi world' } },
+              { data: { content: 'Hello? Or hi?' } },
+            ],
+          });
+
+          const [user, user2] = await createItems({
+            keystone,
+            listKey: 'User',
+            items: [
+              { data: { posts: { connect: ids } } },
+              { data: { posts: { connect: [ids[0]] } } }, // Create a dummy user to make sure we're actually filtering it out
+            ],
+          });
+
+          const { data, errors } = await keystone.executeGraphQL({
             query: `
         query {
           allUsers {
             id
             posts (first: 1, sortBy: content_ASC) {
               id
-              content
             }
           }
         }
@@ -104,38 +118,34 @@ multiAdapterRunners().map(({ runner, adapterName }) =>
 
           expect(errors).toBe(undefined);
           expect(data).toHaveProperty('allUsers.0.posts');
-          expect(data.allUsers[0].posts).toHaveLength(1);
-          expect(data).toMatchObject({
-            allUsers: [
-              {
-                id: user.id,
-                posts: [{ id: ids[0] }],
-              },
-              {
-                id: user2.id,
-                posts: [{ id: ids[0] }],
-              },
-            ],
-          });
+          expect(data.allUsers).toContainEqual({ id: user.id, posts: [ids[0]] });
+          expect(data.allUsers).toContainEqual({ id: user2.id, posts: [ids[0]] });
         })
       );
 
       test(
         'nested to-many relationships can be filtered within AND clause',
-        runner(setupKeystone, async ({ keystone, create }) => {
-          const ids = [];
-
-          ids.push((await create('Post', { content: 'Hello world' })).id);
-          ids.push((await create('Post', { content: 'hi world' })).id);
-          ids.push((await create('Post', { content: 'Hello? Or hi?' })).id);
-
-          const user = await create('User', { posts: ids });
-
-          // Create a dummy user to make sure we're actually filtering it out
-          const user2 = await create('User', { posts: [ids[0]] });
-
-          const { data, errors } = await graphqlRequest({
+        runner(setupKeystone, async ({ keystone }) => {
+          const ids = await createItems({
             keystone,
+            listKey: 'Post',
+            items: [
+              { data: { content: 'Hello world' } },
+              { data: { content: 'hi world' } },
+              { data: { content: 'Hello? Or hi?' } },
+            ],
+          });
+
+          const [user, user2] = await createItems({
+            keystone,
+            listKey: 'User',
+            items: [
+              { data: { posts: { connect: ids } } },
+              { data: { posts: { connect: [ids[0]] } } }, // Create a dummy user to make sure we're actually filtering it out
+            ],
+          });
+
+          const { data, errors } = await keystone.executeGraphQL({
             query: `
         query {
           allUsers {
@@ -147,7 +157,6 @@ multiAdapterRunners().map(({ runner, adapterName }) =>
               ]
             }){
               id
-              content
             }
           }
         }
@@ -156,38 +165,34 @@ multiAdapterRunners().map(({ runner, adapterName }) =>
 
           expect(errors).toBe(undefined);
           expect(data).toHaveProperty('allUsers.0.posts');
-          expect(data.allUsers[0].posts).toHaveLength(1);
-          expect(data).toMatchObject({
-            allUsers: [
-              {
-                id: user.id,
-                posts: [{ id: ids[2] }],
-              },
-              {
-                id: user2.id,
-                posts: [],
-              },
-            ],
-          });
+          expect(data.allUsers).toContainEqual({ id: user.id, posts: [ids[2]] });
+          expect(data.allUsers).toContainEqual({ id: user2.id, posts: [] });
         })
       );
 
       test(
         'nested to-many relationships can be filtered within OR clause',
-        runner(setupKeystone, async ({ keystone, create }) => {
-          const ids = [];
-
-          ids.push((await create('Post', { content: 'Hello world' })).id);
-          ids.push((await create('Post', { content: 'hi world' })).id);
-          ids.push((await create('Post', { content: 'Hello? Or hi?' })).id);
-
-          const user = await create('User', { posts: ids });
-
-          // Create a dummy user to make sure we're actually filtering it out
-          const user2 = await create('User', { posts: [ids[0]] });
-
-          const { data, errors } = await graphqlRequest({
+        runner(setupKeystone, async ({ keystone }) => {
+          const ids = await createItems({
             keystone,
+            listKey: 'Post',
+            items: [
+              { data: { content: 'Hello world' } },
+              { data: { content: 'hi world' } },
+              { data: { content: 'Hello? Or hi?' } },
+            ],
+          });
+
+          const [user, user2] = await createItems({
+            keystone,
+            listKey: 'User',
+            items: [
+              { data: { posts: { connect: ids } } },
+              { data: { posts: { connect: [ids[0]] } } }, // Create a dummy user to make sure we're actually filtering it out
+            ],
+          });
+
+          const { data, errors } = await keystone.executeGraphQL({
             query: `
         query {
           allUsers {
@@ -208,34 +213,23 @@ multiAdapterRunners().map(({ runner, adapterName }) =>
 
           expect(errors).toBe(undefined);
           expect(data).toHaveProperty('allUsers.0.posts');
-          expect(data).toMatchObject({
-            allUsers: [
-              {
-                id: user.id,
-                posts: expect.arrayContaining([
-                  expect.objectContaining({ id: ids[1] }),
-                  expect.objectContaining({ id: ids[2] }),
-                ]),
-              },
-              {
-                id: user2.id,
-                posts: [],
-              },
-            ],
+          expect(data.allUsers).toContainEqual({
+            id: user.id,
+            posts: expect.arrayContaining([
+              expect.objectContaining(ids[1]),
+              expect.objectContaining(ids[2]),
+            ]),
           });
-          // `expect.arrayContaining()` doesn't fail if there are _more_ results
-          // than expected
-          expect(data.allUsers[0].posts).toHaveLength(2);
+          expect(data.allUsers).toContainEqual({ id: user2.id, posts: [] });
         })
       );
 
       test(
         'Filtering out all items by nested field should return []',
-        runner(setupKeystone, async ({ keystone, create }) => {
-          await create('User', {});
+        runner(setupKeystone, async ({ keystone }) => {
+          await createItem({ keystone, listKey: 'User', item: {} });
 
-          const result = await graphqlRequest({
-            keystone,
+          const result = await keystone.executeGraphQL({
             query: `
               query {
                 allUsers(where: { posts_some: { content_contains: "foo" } }) {
@@ -255,20 +249,27 @@ multiAdapterRunners().map(({ runner, adapterName }) =>
     describe('relationship meta filtering', () => {
       test(
         'nested to-many relationships return meta info',
-        runner(setupKeystone, async ({ keystone, create }) => {
-          const ids = [];
-
-          ids.push((await create('Post', { content: 'Hello world' })).id);
-          ids.push((await create('Post', { content: 'hi world' })).id);
-          ids.push((await create('Post', { content: 'Hello? Or hi?' })).id);
-
-          const user = await create('User', { posts: ids });
-
-          // Create a dummy user to make sure we're actually filtering it out
-          const user2 = await create('User', { posts: [ids[0]] });
-
-          const { data, errors } = await graphqlRequest({
+        runner(setupKeystone, async ({ keystone }) => {
+          const ids = await createItems({
             keystone,
+            listKey: 'Post',
+            items: [
+              { data: { content: 'Hello world' } },
+              { data: { content: 'hi world' } },
+              { data: { content: 'Hello? Or hi?' } },
+            ],
+          });
+
+          const [user, user2] = await createItems({
+            keystone,
+            listKey: 'User',
+            items: [
+              { data: { posts: { connect: ids } } },
+              { data: { posts: { connect: [ids[0]] } } }, // Create a dummy user to make sure we're actually filtering it out
+            ],
+          });
+
+          const { data, errors } = await keystone.executeGraphQL({
             query: `
         query {
           allUsers {
@@ -284,37 +285,34 @@ multiAdapterRunners().map(({ runner, adapterName }) =>
           expect(errors).toBe(undefined);
           expect(data.allUsers).toHaveLength(2);
           expect(data).toHaveProperty('allUsers.0._postsMeta');
-          expect(data).toMatchObject({
-            allUsers: [
-              {
-                id: user.id,
-                _postsMeta: { count: 3 },
-              },
-              {
-                id: user2.id,
-                _postsMeta: { count: 1 },
-              },
-            ],
-          });
+          expect(data.allUsers).toContainEqual({ id: user.id, _postsMeta: { count: 3 } });
+          expect(data.allUsers).toContainEqual({ id: user2.id, _postsMeta: { count: 1 } });
         })
       );
 
       test(
         'nested to-many relationship meta can be filtered',
-        runner(setupKeystone, async ({ keystone, create }) => {
-          const ids = [];
-
-          ids.push((await create('Post', { content: 'Hello world' })).id);
-          ids.push((await create('Post', { content: 'hi world' })).id);
-          ids.push((await create('Post', { content: 'Hello? Or hi?' })).id);
-
-          const user = await create('User', { posts: ids });
-
-          // Create a dummy user to make sure we're actually filtering it out
-          const user2 = await create('User', { posts: [ids[0]] });
-
-          const { data, errors } = await graphqlRequest({
+        runner(setupKeystone, async ({ keystone }) => {
+          const ids = await createItems({
             keystone,
+            listKey: 'Post',
+            items: [
+              { data: { content: 'Hello world' } },
+              { data: { content: 'hi world' } },
+              { data: { content: 'Hello? Or hi?' } },
+            ],
+          });
+
+          const [user, user2] = await createItems({
+            keystone,
+            listKey: 'User',
+            items: [
+              { data: { posts: { connect: ids } } },
+              { data: { posts: { connect: [ids[0]] } } }, // Create a dummy user to make sure we're actually filtering it out
+            ],
+          });
+
+          const { data, errors } = await keystone.executeGraphQL({
             query: `
         query {
           allUsers {
@@ -332,37 +330,34 @@ multiAdapterRunners().map(({ runner, adapterName }) =>
           expect(errors).toBe(undefined);
           expect(data.allUsers).toHaveLength(2);
           expect(data).toHaveProperty('allUsers.0._postsMeta');
-          expect(data).toMatchObject({
-            allUsers: [
-              {
-                id: user.id,
-                _postsMeta: { count: 2 },
-              },
-              {
-                id: user2.id,
-                _postsMeta: { count: 0 },
-              },
-            ],
-          });
+          expect(data.allUsers).toContainEqual({ id: user.id, _postsMeta: { count: 2 } });
+          expect(data.allUsers).toContainEqual({ id: user2.id, _postsMeta: { count: 0 } });
         })
       );
 
       test(
         'nested to-many relationship meta can be limited',
-        runner(setupKeystone, async ({ keystone, create }) => {
-          const ids = [];
-
-          ids.push((await create('Post', { content: 'Hello world' })).id);
-          ids.push((await create('Post', { content: 'hi world' })).id);
-          ids.push((await create('Post', { content: 'Hello? Or hi?' })).id);
-
-          const user = await create('User', { posts: ids });
-
-          // Create a dummy user to make sure we're actually filtering it out
-          const user2 = await create('User', { posts: [ids[0]] });
-
-          const { data, errors } = await graphqlRequest({
+        runner(setupKeystone, async ({ keystone }) => {
+          const ids = await createItems({
             keystone,
+            listKey: 'Post',
+            items: [
+              { data: { content: 'Hello world' } },
+              { data: { content: 'hi world' } },
+              { data: { content: 'Hello? Or hi?' } },
+            ],
+          });
+
+          const [user, user2] = await createItems({
+            keystone,
+            listKey: 'User',
+            items: [
+              { data: { posts: { connect: ids } } },
+              { data: { posts: { connect: [ids[0]] } } }, // Create a dummy user to make sure we're actually filtering it out
+            ],
+          });
+
+          const { data, errors } = await keystone.executeGraphQL({
             query: `
         query {
           allUsers {
@@ -378,37 +373,34 @@ multiAdapterRunners().map(({ runner, adapterName }) =>
           expect(errors).toBe(undefined);
           expect(data).toHaveProperty('allUsers.0._postsMeta');
           expect(data.allUsers).toHaveLength(2);
-          expect(data).toMatchObject({
-            allUsers: [
-              {
-                id: user.id,
-                _postsMeta: { count: 1 },
-              },
-              {
-                id: user2.id,
-                _postsMeta: { count: 1 },
-              },
-            ],
-          });
+          expect(data.allUsers).toContainEqual({ id: user.id, _postsMeta: { count: 1 } });
+          expect(data.allUsers).toContainEqual({ id: user2.id, _postsMeta: { count: 1 } });
         })
       );
 
       test(
         'nested to-many relationship meta can be filtered within AND clause',
-        runner(setupKeystone, async ({ keystone, create }) => {
-          const ids = [];
-
-          ids.push((await create('Post', { content: 'Hello world' })).id);
-          ids.push((await create('Post', { content: 'hi world' })).id);
-          ids.push((await create('Post', { content: 'Hello? Or hi?' })).id);
-
-          const user = await create('User', { posts: ids });
-
-          // Create a dummy user to make sure we're actually filtering it out
-          const user2 = await create('User', { posts: [ids[0]] });
-
-          const { data, errors } = await graphqlRequest({
+        runner(setupKeystone, async ({ keystone }) => {
+          const ids = await createItems({
             keystone,
+            listKey: 'Post',
+            items: [
+              { data: { content: 'Hello world' } },
+              { data: { content: 'hi world' } },
+              { data: { content: 'Hello? Or hi?' } },
+            ],
+          });
+
+          const [user, user2] = await createItems({
+            keystone,
+            listKey: 'User',
+            items: [
+              { data: { posts: { connect: ids } } },
+              { data: { posts: { connect: [ids[0]] } } }, // Create a dummy user to make sure we're actually filtering it out
+            ],
+          });
+
+          const { data, errors } = await keystone.executeGraphQL({
             query: `
         query {
           allUsers {
@@ -429,37 +421,34 @@ multiAdapterRunners().map(({ runner, adapterName }) =>
           expect(errors).toBe(undefined);
           expect(data.allUsers).toHaveLength(2);
           expect(data).toHaveProperty('allUsers.0._postsMeta');
-          expect(data).toMatchObject({
-            allUsers: [
-              {
-                id: user.id,
-                _postsMeta: { count: 1 },
-              },
-              {
-                id: user2.id,
-                _postsMeta: { count: 0 },
-              },
-            ],
-          });
+          expect(data.allUsers).toContainEqual({ id: user.id, _postsMeta: { count: 1 } });
+          expect(data.allUsers).toContainEqual({ id: user2.id, _postsMeta: { count: 0 } });
         })
       );
 
       test(
         'nested to-many relationship meta can be filtered within OR clause',
-        runner(setupKeystone, async ({ keystone, create }) => {
-          const ids = [];
-
-          ids.push((await create('Post', { content: 'Hello world' })).id);
-          ids.push((await create('Post', { content: 'hi world' })).id);
-          ids.push((await create('Post', { content: 'Hello? Or hi?' })).id);
-
-          const user = await create('User', { posts: ids });
-
-          // Create a dummy user to make sure we're actually filtering it out
-          const user2 = await create('User', { posts: [ids[0]] });
-
-          const { data, errors } = await graphqlRequest({
+        runner(setupKeystone, async ({ keystone }) => {
+          const ids = await createItems({
             keystone,
+            listKey: 'Post',
+            items: [
+              { data: { content: 'Hello world' } },
+              { data: { content: 'hi world' } },
+              { data: { content: 'Hello? Or hi?' } },
+            ],
+          });
+
+          const [user, user2] = await createItems({
+            keystone,
+            listKey: 'User',
+            items: [
+              { data: { posts: { connect: ids } } },
+              { data: { posts: { connect: [ids[0]] } } }, // Create a dummy user to make sure we're actually filtering it out
+            ],
+          });
+
+          const { data, errors } = await keystone.executeGraphQL({
             query: `
         query {
           allUsers {
@@ -480,18 +469,8 @@ multiAdapterRunners().map(({ runner, adapterName }) =>
           expect(errors).toBe(undefined);
           expect(data.allUsers).toHaveLength(2);
           expect(data).toHaveProperty('allUsers.0._postsMeta');
-          expect(data).toMatchObject({
-            allUsers: [
-              {
-                id: user.id,
-                _postsMeta: { count: 2 },
-              },
-              {
-                id: user2.id,
-                _postsMeta: { count: 0 },
-              },
-            ],
-          });
+          expect(data.allUsers).toContainEqual({ id: user.id, _postsMeta: { count: 2 } });
+          expect(data.allUsers).toContainEqual({ id: user2.id, _postsMeta: { count: 0 } });
         })
       );
     });