+++ packages/fields-content/src/blocks/paragraph.js
@@ -1,0 +1,11 @@
+import { resolveView } from '../resolve-view';
+import { Block } from '../Block';
+
+export default class ParagraphBlock extends Block {
+  get type() {
+    return 'paragraph';
+  }
+  getAdminViews() {
+    return [resolveView('views/editor/blocks/paragraph')];
+  }
+}