+++ packages/fields-mongoid/src/test-fixtures.js
@@ -1,0 +1,7 @@
+import { MongoId } from './index';
+
+export const name = 'MongoId';
+export { MongoId as type };
+export const exampleValue = '"123456781234567812345678"';
+export const exampleValue2 = '"123456781234567812345679"';
+export const supportsUnique = true;