+++ packages/fields/src/types/CalendarDay/Implementation.js
@@ -33,6 +33,10 @@ export class CalendarDay extends Implementation {
     this.isOrderable = true;
   }
 
+  get _supportsUnique() {
+    return true;
+  }
+
   gqlOutputFields() {
     return [`${this.path}: String`];
   }