+++ packages/fields/src/types/Unsplash/UnsplashBlock.js
@@ -1,7 +1,7 @@
 import pluralize from 'pluralize';
 import { importView } from '@keystonejs/build-field-types';
 
-import { Block } from '@keystonejs/field-content/Block';
+import { Block } from '@keystonejs/fields-content/Block';
 import Unsplash from './';
 import RelationshipType from '../Relationship';