+++ packages/fields/src/types/Unsplash/views/blocks/unsplash-image.js
@@ -2,11 +2,10 @@
 import { jsx } from '@emotion/core';
 import { Fragment, useState, useCallback, createContext, useContext } from 'react';
 import pluralize from 'pluralize';
-import { BlockMenuItem } from '@keystonejs/field-content/block-components';
+import { BlockMenuItem } from '@keystonejs/fields-content/block-components';
 import { Input } from '@arch-ui/input';
 
-import { useLazyQuery } from '@apollo/react-hooks';
-import gql from 'graphql-tag';
+import { gql, useLazyQuery } from '@apollo/client';
 
 export let type = 'unsplashImage';