+++ packages/oembed-adapters/src/iframely.js
@@ -1,12 +1,17 @@
+import path from 'path';
 import fetch from 'node-fetch';
 import crypto from 'crypto';
-import { importView } from '@keystonejs/build-field-types';
 
 const VALID_URL = /^https?:\/\//i;
 const IS_MD5 = /[a-f0-9]{32}/i;
 
+const previewComponent = path.join(
+  path.dirname(require.resolve('@keystonejs/oembed-adapters/package.json')),
+  'views/preview'
+);
+
 export class IframelyOEmbedAdapter {
-  constructor({ apiKey } = {}) {
+  constructor({ apiKey, parameters } = {}) {
     if (!apiKey) {
       throw new Error('Must provide an apiKey to IFramely OEmbed Adapter');
     }
@@ -21,6 +26,8 @@ export class IframelyOEmbedAdapter {
         .update(apiKey)
         .digest('hex');
     }
+
+    this.requestParams = parameters || {};
   }
 
   /**
@@ -40,14 +47,13 @@ export class IframelyOEmbedAdapter {
     }
 
     const params = Object.entries({
-      // Force all `html` to be returned as an iFramely iFrame:
       // https://iframely.com/docs/iframes
-      iframe: '1',
-      // Assume the client will load the script themselves. This is important
       // for React apps.
       // https://iframely.com/docs/reactjs
-      omit_script: '1',
       // Allow overwriting most parameters
+      iframe: '1',
+      omit_script: '1',
+      ...this.requestParams,
       ...parameters,
       // We're using the MD5 hashed key:
       // https://iframely.com/docs/allow-origins
@@ -58,12 +64,12 @@ export class IframelyOEmbedAdapter {
   }
 
   getAdminViews() {
-    return [importView('./views/preview')];
+    return [previewComponent];
   }
 
   getViewOptions() {
     return {
-      previewComponent: importView('./views/preview'),
+      previewComponent,
       // NOTE: This is the md5'd API key from the constructor, which is ok to
       // put on the client according to the docs
       clientApiKey: this.apiKey,