+++ test-projects/basic/index.js
@@ -10,7 +10,6 @@ const {
   CalendarDay,
   CloudinaryImage,
   DateTime,
-  Color,
   Url,
   Decimal,
   OEmbed,
@@ -18,10 +17,11 @@ const {
   Unsplash,
   Virtual,
 } = require('@keystonejs/fields');
-const { Content } = require('@keystonejs/field-content');
+const { Content } = require('@keystonejs/fields-content');
 const { CloudinaryAdapter, LocalFileAdapter } = require('@keystonejs/file-adapters');
 const { Markdown } = require('@keystonejs/fields-markdown');
 const { Wysiwyg } = require('@keystonejs/fields-wysiwyg-tinymce');
+const { Color } = require('@keystonejs/fields-color');
 const { GraphQLApp } = require('@keystonejs/app-graphql');
 const { AdminUIApp } = require('@keystonejs/app-admin-ui');
 const { StaticApp } = require('@keystonejs/app-static');